AxionApp/requirements.txt
asgiref==3.2.10
autopep8==1.5.6
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-import-export==2.5.0
django-jquery==3.1.0
django-static-jquery==2.1.4
et-xmlfile==1.1.0
gunicorn==20.0.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
static3==0.7.0
tablib==3.0.0
toml==0.10.2
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0